Rename all modules
- Drop pureport_ prefix from all modules and doc_fragments
- Rename module _facts to _info
- Symlink all old names for backwards compatability
- Update README
