+++ plugins/doc_fragments/resolve_existing.py
