+++ plugins/modules/access_token_info.py
@@ -1,0 +1,98 @@
+#!/usr/bin/python
+
+ANSIBLE_METADATA = {
+    'metadata_version': '1.1',
+    'status': ['preview'],
+    'supported_by': 'Pureport'
+}
+
+DOCUMENTATION = '''
+---
+module: access_token_info
+short_description: Retrieve an access token to use with the Pureport API
+description:
+    - "Retrieve an access token to use with the Pureport API"
+version_added: "2.8"
+requirements: [ pureport-client ]
+author: Matt Traynham (@mtraynham)
+options:
+    api_base_url:
+        description:
+            - The host url for the Pureport API.
+        required: false
+        type: str
+    api_key:
+        description:
+            - The pre-configured API Key for a Pureport Account.
+        required: true
+        type: str
+    api_secret:
+        description:
+            - The pre-configured API Secret for a Pureport Account.
+        required: true
+        type: str
+'''
+
+EXAMPLES = '''
+- name: Retrieve the access token for an api key and secret
+  access_token_info:
+    api_key: XXXXXXXXXXXXX
+    api_secret: XXXXXXXXXXXXXXXXX
+  register: result   # Registers result.access_token
+
+- name: Set the access token as a fact
+  set_fact:
+    access_token: result.access_token
+'''
+
+RETURN = '''
+access_token:
+    description:
+        - An access token that can be used with other Pureport facts.
+    returned: success
+    type: str
+'''
+
+from ansible.module_utils.basic import AnsibleModule
+from traceback import format_exc
+
+try:
+    from pureport.api.client import Client
+    from pureport.exception.api import ClientHttpException
+    HAS_PUREPORT_CLIENT = True
+except ImportError:
+    HAS_PUREPORT_CLIENT = False
+    Client = None
+    ClientHttpException = None
+
+
+def get_access_token(module):
+    """
+    Get the access token
+    :param AnsibleModule module: the ansible module
+    """
+    if not HAS_PUREPORT_CLIENT:
+        module.fail_json(msg='pureport-client required for this module')
+    client = Client(module.params.get('api_base_url'))
+    try:
+        return client.login(module.params.get('api_key'), module.params.get('api_secret'))
+    except ClientHttpException as e:
+        module.fail_json(msg=e.response.text, exception=format_exc())
+
+
+def main():
+    argument_spec = dict(
+        api_base_url=dict(type='str'),
+        api_key=dict(type='str', required=True),
+        api_secret=dict(type='str', required=True, no_log=True)
+    )
+    mutually_exclusive = []
+    module = AnsibleModule(
+        argument_spec=argument_spec,
+        mutually_exclusive=mutually_exclusive
+    )
+    module.exit_json(access_token=get_access_token(module))
+
+
+if __name__ == '__main__':
+    main()
